+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Prism-Wiki now supports Node.js environments
 - Auto-linking for the `src` attribute of `<templatestyles>` extension tags in Wikitext code blocks
+- Auto-linking for Scribunto module invocations in Wikitext code blocks
 
 **Fixed**